src - Move connection classes to new module
